final edits
I just went through and changed the wording on a few things. I added a
bit the the final corollary, to make it clearer. I also made all the
locations have bars over them to indicate that they are vectors. We
could make them bold instead if you think that is cleaner but I would
like to have some for of indication in the notation that they represent
a vector.